+++ packages/expo-notifications/android/src/test/java/expo/modules/notifications/NotificationDataTest.kt
@@ -3,6 +3,8 @@ package expo.modules.notifications
 import expo.modules.notifications.notifications.model.NotificationData
 import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -17,7 +19,8 @@ class NotificationDataTest {
       "title" to "My Title",
       "message" to "My Message",
       "sound" to "default",
-      "body" to "My Body",
+      "body" to "{\"key\":\"value\"}",
+      "color" to "#FF0000",
       "categoryId" to "myCategory",
       "subtitle" to "My Subtitle",
       "badge" to "1"
@@ -26,16 +29,19 @@ class NotificationDataTest {
     assertEquals("My Title", notificationData.title)
     assertEquals("My Message", notificationData.message)
     assertEquals("default", notificationData.sound)
-    assertEquals("My Body", notificationData.body)
+    assertNotNull(notificationData.body)
     assertEquals("myCategory", notificationData.categoryId)
     assertEquals("My Subtitle", notificationData.subText)
+    assertEquals("#FF0000", notificationData.color)
     assertEquals(1, notificationData.badge)
+    assertFalse(notificationData.shouldPlayDefaultSound)
   }
 
   @Test
   fun testVibrationPattern() {
     val notificationData = NotificationData(mapOf("vibrate" to "[100, 200, 300]"))
     assertArrayEquals(longArrayOf(100, 200, 300), notificationData.vibrationPattern)
+    assertFalse(notificationData.shouldUseDefaultVibrationPattern)
   }
 
   @Test
@@ -55,5 +61,7 @@ class NotificationDataTest {
     assertEquals(true, notificationData.shouldUseDefaultVibrationPattern)
     assertEquals(false, notificationData.autoDismiss)
     assertEquals(true, notificationData.sticky)
+
+    assertEquals(true, notificationData.shouldPlayDefaultSound)
   }
 }